services/web/requirements.txt
eventlet==0.35.2
flask-login==0.6.3
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-wtf==1.2.1
flask==3.0.2
gunicorn==21.2.0
intelhex==2.3.0
pycryptodome
python-dotenv==1.0.0
pytz
requests==2.31.0
wtforms_sqlalchemy==0.4.1